grabbing files from local
