+++ resource_estimation/ftqc/__init__.py
@@ -57,6 +57,7 @@ from resource_estimation.ftqc.stim_functions import (
     cultivate,
     load_saved_cost,
 )
+from resource_estimation.ftqc.distil import distil_15_to_1
 
 __all__ = [
     "Architecture",
@@ -95,4 +96,5 @@ __all__ = [
     "teleport_S",
     "teleport_T",
     "validate_ops",
+    "distil_15_to_1"
 ]
