+++ resource_estimation/ftqc/architecture.py
@@ -11,8 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from __future__ import annotations
-
 import abc
 import json
 from collections import Counter
@@ -22,12 +20,12 @@ from pathlib import Path
 
 import cirq
 import numpy as np
-from cirq.ops.measurement_gate import MeasurementGate
-from cirq.ops.permutation_gate import QubitPermutationGate
-from cirq.ops.raw_types import Gate, Operation
 from cirq_superstaq.ops.qubit_gates import ParallelRGate
 
-from resource_estimation.ftqc.stim_functions import cultivate
+from resource_estimation.compile_ftqc import add_moves
+from resource_estimation.distil import distil_15_to_1
+from resource_estimation.estimate import ResourceEstimator
+from resource_estimation.stim_functions import cultivate
 
 from . import lattice_surgery_primitives as lsp
 
@@ -55,8 +53,7 @@ def _merge_cost(
     k: int,
     smooth: bool,
 ) -> dict[str, dict[cirq.Gate, int]]:
-    """
-    Calculates the resources required to implement the merge operation.
+    """Calculates the resources required to implement the merge operation.
     d is the code distance, k is the number of patches being merged, and smooth indicates if patches are being joined at the Z or X boundary
     """
     assert k >= 2
@@ -117,8 +114,7 @@ def _syndrome_extract_cost(
     num_logical_qubits: int,
     d: int,
 ) -> dict[str, dict[cirq.Gate, int]]:
-    """
-    Calculates the cost of syndrome extraction in terms of physical gates
+    """Calculates the cost of syndrome extraction in terms of physical gates
     """
     # This is how SE should look...
     # ...for a (full) X stabilizer
@@ -160,8 +156,7 @@ def _syndrome_extract_cost(
 
 @lru_cache(maxsize=128)
 def _split_cost(smooth: bool, d: int) -> dict[str, dict[cirq.Gate, int]]:
-    """
-    Calculates cost to perform a split operation
+    """Calculates cost to perform a split operation
     Split operations can always be absorbed into a following moment, so the moment cost is null
     """
     if smooth:
@@ -179,11 +174,10 @@ def _split_cost(smooth: bool, d: int) -> dict[str, dict[cirq.Gate, int]]:
 
 
 class Architecture(abc.ABC):
-    """
-    Class for representing device architectures.
+    """Class for representing device architectures.
 
-        Generally, only subclasses of this class should be used.
-        Comes preloaded with many primitive costs that are shared among the current set of subclasses.
+    Generally, only subclasses of this class should be used.
+    Comes preloaded with many primitive costs that are shared among the current set of subclasses.
     """
 
     def __init__(
@@ -214,7 +208,7 @@ class Architecture(abc.ABC):
     ### Class Methods ###
     # TODO: Deprecate these
     @classmethod
-    def from_dict(cls, d: dict) -> DefaultLattice | DefaultMovement:
+    def from_dict(cls, d: dict):
         movement = d["movement"]
         if movement:
             base_arc = DefaultMovement(
@@ -256,7 +250,7 @@ class Architecture(abc.ABC):
         return base_arc
 
     @classmethod
-    def from_json(cls, fp: str | Path) -> DefaultLattice | DefaultMovement:
+    def from_json(cls, fp: str | Path):
         with open(fp) as f:
             input_dict = json.load(f)
         return cls.from_dict(input_dict)
@@ -298,12 +292,10 @@ class Architecture(abc.ABC):
 
     @cached_property
     def _cultivate_y_cost(self):
-        """
-        Cost estimate for measuring a surface code patch in the Y basis. Measuring in the Y basis facilitates gate teleportation the same way that cultivating T does.
+        """Cost estimate for measuring a surface code patch in the Y basis. Measuring in the Y basis facilitates gate teleportation the same way that cultivating T does.
         The procedure is based on [Inplace access to the Surface Code Y Basis](https://arxiv.org/pdf/2302.07395v2).
         The cost estimates were generated by looking carefully at circuits produced from https://doi.org/10.5281/zenodo.7487893.
         """
-
         single_qubit_moments = 5
         reset_moments = 1
         cz_moments = 5
@@ -371,11 +363,10 @@ class Architecture(abc.ABC):
         return {"op_time": op_time, "gate_cost": gate_cost, "moment_cost": moment_cost}
 
     @property
-    def rounds(self) -> int:
+    def rounds(self):
         if self.syndrome_rounds is None:
             return self.d
-        else:
-            return self.syndrome_rounds
+        return self.syndrome_rounds
 
     @property
     def phys_gate_times(self) -> dict[cirq.Gate, float]:
@@ -401,10 +392,9 @@ class Architecture(abc.ABC):
         theta = op.gate.theta
         if np.isclose(theta, np.pi / 2):
             return self._cultivate_y_cost
-        elif np.isclose(theta, np.pi / 4):
+        if np.isclose(theta, np.pi / 4):
             return self._cultivate_t_cost
-        else:
-            raise ValueError(f"Cultivation cost is not defined for angle: {theta}")
+        raise ValueError(f"Cultivation cost is not defined for angle: {theta}")
 
     def syndrome_extract_cost(self, op: cirq.Operation) -> dict:
         cost_dict = _syndrome_extract_cost(
@@ -441,7 +431,7 @@ class Architecture(abc.ABC):
         return self._h_cost
 
     ### Extra Methods ###
-    def __post_init__(self) -> None:
+    def __post_init__(self):
         # Initialize with all shared Primitives then add special ones later
         self.op_cost = {
             lsp.Cultivate: self.cultivate_cost,
@@ -466,8 +456,7 @@ class Architecture(abc.ABC):
 
 
 class DefaultLattice(Architecture):
-    """
-    The subclass used to represent Dual Species without movement
+    """The subclass used to represent Dual Species without movement
     It uses lattice surgery operations assumes no correlated decoding
     """
 
@@ -475,8 +464,8 @@ class DefaultLattice(Architecture):
         self,
         idling: bool = True,
         post_op_correction: bool = True,
-        d: int = 7,
-        cultivation_repetition: int = 1,
+        d=7,
+        cultivation_repetition=1,
         cultivation_fault_distance: int = 3,
         syndrome_rounds=None,
     ) -> None:
@@ -549,7 +538,7 @@ class DefaultLattice(Architecture):
         return {"op_time": op_time, "gate_cost": gate_cost, "moment_cost": moment_cost}
 
     @cached_property
-    def _cultivate_t_cost(self) -> dict[str, dict[Gate, int] | float]:
+    def _cultivate_t_cost(self):
         # fold should always be false here
         base_cultivation_cost = cultivate(
             dsurface=self.d, fold=self.fold_cultiv, fault_distance=self.cultivation_fault_distance
@@ -579,8 +568,7 @@ class DefaultLattice(Architecture):
 
 
 class DefaultMovement(Architecture):
-    """
-    Class representing the set of Primitives available with access to movement for transversal operations.
+    """Class representing the set of Primitives available with access to movement for transversal operations.
     This default version assumes a single species of neutral atom qubits using a zoned architecture.
     """
 
@@ -589,10 +577,11 @@ class DefaultMovement(Architecture):
         idling: bool = False,
         post_op_correction: bool = True,
         d: int = 7,
-        fold_cultiv: bool = False,
-        cultivation_repetition: int = 1,
+        fold_cultiv=False,
+        cultivation_repetition=1,
+        distillation_repetition=1,
         cultivation_fault_distance: int = 3,
-        syndrome_rounds: int = 1,
+        syndrome_rounds=1,
     ) -> None:
         super().__init__(
             idling=idling,
@@ -604,9 +593,11 @@ class DefaultMovement(Architecture):
             syndrome_rounds=syndrome_rounds,
             fold_cultiv=fold_cultiv,
         )
+        self.distillation_repetition = distillation_repetition
         self._primitives = cirq.Gateset(
             *[
                 lsp.Cultivate,
+                lsp.Distil,
                 lsp.SyndromeExtract,
                 lsp.ErrorCorrect,
                 lsp.Move,
@@ -628,7 +619,7 @@ class DefaultMovement(Architecture):
     def cnot_cost(self, op: cirq.Operation) -> dict:
         return self._cnot_cost
 
-    def syndrome_extract_cost(self, op: Operation):
+    def syndrome_extract_cost(self, op):
         # Build from the base cost of Syndrome Extraction by adding movement penalties CZ and Measurement moments
         base_cost = super().syndrome_extract_cost(op).copy()
         moment_cost = base_cost["moment_cost"]
@@ -674,8 +665,7 @@ class DefaultMovement(Architecture):
 
     @cached_property
     def _s_cost(self) -> dict:
-        """
-        Resources the fold transversal S gate from https://arxiv.org/pdf/2412.01391.
+        """Resources the fold transversal S gate from https://arxiv.org/pdf/2412.01391.
         It looks like one Syndrome Extraction round with some CNOT gates across the main diagonal, as well as some physical S/Sdg gates.
         """
         # precompute syndrome extraction cost
@@ -696,9 +686,8 @@ class DefaultMovement(Architecture):
         op_time = self.total_time(moment_cost_dict=moment_cost)
         return {"op_time": op_time, "gate_cost": gate_cost, "moment_cost": moment_cost}
 
-    def move_cost(self, op) -> dict[str, dict[type[QubitPermutationGate], int] | float]:
-        """
-        Method to handle both types of movement
+    def move_cost(self, op):
+        """Method to handle both types of movement
         The maximum move time should be 500us, which corresponds to travelling to a zone
         Everything else should be penalized by distance away up to a distance of 500us
         This reference says something about .55um/us (https://www.nature.com/articles/s41586-022-04592-6.pdf)
@@ -723,7 +712,7 @@ class DefaultMovement(Architecture):
         return {"op_time": op_time, "gate_cost": gate_cost, "moment_cost": moment_cost}
 
     @cached_property
-    def _cultivate_t_cost(self) -> dict[str, dict[Gate, int] | float]:
+    def _cultivate_t_cost(self):
         base_cultivation_cost = cultivate(
             dsurface=self.d, fold=self.fold_cultiv, fault_distance=self.cultivation_fault_distance
         ).copy()
@@ -760,11 +749,38 @@ class DefaultMovement(Architecture):
         new_time = self.total_time(new_moment_cost)
         return {"op_time": new_time, "gate_cost": new_gate_cost, "moment_cost": new_moment_cost}
 
+    def distil_t_cost(self, op: cirq.Operation) -> dict:
+        return self._distil_t_cost
+
+    @cached_property
+    def _distil_t_cost(self) -> dict:
+        """Cost to get a T state using 15-to-1 distillation
+        """
+        mapped_circuit = distil_15_to_1()
+        with_moves = add_moves(
+            mapped_circuit,
+            zone_ops=self.zone_ops if self.zone_ops is not None else cirq.Gateset(),
+            alley_ops=self.alley_ops if self.alley_ops is not None else cirq.Gateset(),
+        )
+        estimator = ResourceEstimator(self)
+        rep_time = estimator.parallel_circuit_time(with_moves)
+        rep_moments = estimator.parallel_circuit_cost(with_moves)
+        rep_gates = estimator.serial_circuit_cost(with_moves)
+        op_time = rep_time * self.distillation_repetition
+        moment_cost = Counter(
+            {key: val * self.distillation_repetition for key, val in rep_moments.items()}
+        )
+        gate_cost = Counter(
+            {key: val * self.distillation_repetition for key, val in rep_gates.items()}
+        )
+        return {"op_time": op_time, "moment_cost": moment_cost, "gate_cost": gate_cost}
+
     def __post_init__(self) -> None:
         super().__post_init__()
         self.op_cost[type(cirq.CNOT)] = self.cnot_cost
         self.op_cost[type(cirq.S)] = self.s_cost
         self.op_cost[lsp.Move] = self.move_cost
+        self.op_cost[lsp.Distil] = self.distil_t_cost
 
     @property
     def __name__(self) -> str:
@@ -772,8 +788,7 @@ class DefaultMovement(Architecture):
 
 
 class DualSpeciesMovement(DefaultMovement):
-    """
-    Architecture that gets the best of both worlds.
+    """Architecture that gets the best of both worlds.
     Atoms of different species can be shuttled along alleyways to get close enough to entangle
     Inplace entanglement and readout are achieved via dual species (and possibly hiding beams)
     CZ gates take place between nearest neighbor physical qubits
@@ -798,9 +813,8 @@ class DualSpeciesMovement(DefaultMovement):
 
     # Cultivate from Lattice Surgery
     @cached_property
-    def _cultivate_t_cost(self) -> dict[str, dict[Gate, int] | float]:
-        """
-        Cached property for the cultivation circuit having the relevant parameters: code distance (d) and movement
+    def _cultivate_t_cost(self):
+        """Cached property for the cultivation circuit having the relevant parameters: code distance (d) and movement
         Values are multiplied by the repeat factor for the architecture instance
         """
         base_cultivation_cost = cultivate(
@@ -826,20 +840,19 @@ class DualSpeciesMovement(DefaultMovement):
 
     # Measurement from Lattice Surgery
     @cached_property
-    def _measure_cost(self) -> dict[str, dict[type[MeasurementGate], int] | float]:
+    def _measure_cost(self):
         gate_cost = {cirq.MeasurementGate: self.patch.num_data_qubits}
         moment_cost = {cirq.MeasurementGate: 1}
         op_time = self.total_time(moment_cost_dict=moment_cost)
         return {"op_time": op_time, "gate_cost": gate_cost, "moment_cost": moment_cost}
 
     @property
-    def __name__(self) -> str:
+    def __name__(self):
         return "DualSpeciesMovement"
 
 
 class MeasureZonesOnly(DefaultMovement):
-    """
-    A movement-based Architecture with a Measurement Zone
+    """A movement-based Architecture with a Measurement Zone
     Atoms can be shuttled along alleyways to get close to each other
     Inplace entanglement is enabled through the use of hiding, avoiding the Interaction Zone
     CZ gates can take place with nearest-neighbor physical qubits
@@ -857,8 +870,7 @@ class MeasureZonesOnly(DefaultMovement):
     # For now, I am going with option a), which is the same as DefaultMovement
 
     def syndrome_extract_cost(self, op: cirq.Operation) -> dict:
-        """
-        Uses lattice surgery Syndrome Extraction but adds moves associated with the measurements.
+        """Uses lattice surgery Syndrome Extraction but adds moves associated with the measurements.
         Since this class is a Movement architecture, its rounds should be low, in accordance with the promise of correlated decoding.
         """
         base_cost = _syndrome_extract_cost(
@@ -872,7 +884,7 @@ class MeasureZonesOnly(DefaultMovement):
         return {"moment_cost": moment_cost, "gate_cost": gate_cost, "op_time": op_time}
 
     @cached_property
-    def _cultivate_t_cost(self) -> dict[str, dict[Gate, int] | float]:
+    def _cultivate_t_cost(self):
         base_cultivation_cost = cultivate(
             dsurface=self.d, fold=self.fold_cultiv, fault_distance=self.cultivation_fault_distance
         ).copy()
@@ -917,8 +929,7 @@ class MeasureZonesOnly(DefaultMovement):
 
 
 class Superconductor(DefaultLattice):
-    """
-    Class to serve as a proxy for superconducting architectures.
+    """Class to serve as a proxy for superconducting architectures.
     It features a gateset composed of CZ + 1Q
     It's main feature is its fast gate speeds compared to all other architectures
     """
@@ -927,8 +938,8 @@ class Superconductor(DefaultLattice):
         self,
         idling: bool = True,
         post_op_correction: bool = True,
-        d: int = 7,
-        cultivation_repetition: int = 1,
+        d=7,
+        cultivation_repetition=1,
         cultivation_fault_distance: int = 3,
         syndrome_rounds=None,
     ) -> None:
@@ -964,8 +975,7 @@ class Superconductor(DefaultLattice):
 
 # Deprecated by removing GR gates
 def convert_globals_to_phasedxz(architecture: Architecture, cost_with_globals: dict) -> dict:
-    """
-    Converts costs defined by GR and Rz into PhasedXZ by removing GR and replacing Rz with PhasedXZ to represent arbitrary single qubit rotations
+    """Converts costs defined by GR and Rz into PhasedXZ by removing GR and replacing Rz with PhasedXZ to represent arbitrary single qubit rotations
     """
     gate_cost = cost_with_globals["gate_cost"].copy()
     if ParallelRGate in gate_cost:
