+++ resource_estimation/ftqc/architecture_test.py
@@ -12,40 +12,37 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import json
+import os
 from collections import Counter
 from math import ceil, pi
-from pathlib import Path
 
 import cirq
 import numpy as np
 import pytest
 from cirq_superstaq import ParallelRGate
 
-import resource_estimation.ftqc.architecture as arch
-import resource_estimation.ftqc.estimate as est
-import resource_estimation.ftqc.lattice_surgery_primitives as lsp
-from resource_estimation.ftqc.architecture import DefaultLattice, DefaultMovement
-from resource_estimation.ftqc.stim_functions import cultivate, load_saved_cost
-
-DATA_DIR = Path(__file__).resolve().parents[2] / "data"
+import resource_estimation.architecture as arch
+import resource_estimation.estimate as est
+import resource_estimation.lattice_surgery_primitives as lsp
+from resource_estimation.stim_functions import cultivate, load_saved_cost
 
 
 @pytest.fixture
-def lattice_architecture() -> DefaultLattice:
+def lattice_architecture():
     return arch.DefaultLattice()
 
 
 @pytest.fixture
-def movement_architecture() -> DefaultMovement:
+def movement_architecture():
     return arch.DefaultMovement()
 
 
-def test_architecture_exceptions(lattice_architecture, movement_architecture) -> None:
+def test_architecture_exceptions(lattice_architecture, movement_architecture):
     with pytest.raises(ValueError, match="Cultivation cost"):
         _ = lattice_architecture.cultivate_cost(lsp.Cultivate(1).on(cirq.GridQubit(0, 0)))
 
 
-def test_inplace_exact(lattice_architecture: arch.DefaultLattice) -> None:
+def test_inplace_exact(lattice_architecture: arch.DefaultLattice):
     # TODO: Brainstorm a better way to test this feature
     actual_op_cost = lattice_architecture.cultivate_cost(
         lsp.Cultivate(pi / 2).on(cirq.GridQubit(0, 0))
@@ -74,14 +71,14 @@ def test_inplace_exact(lattice_architecture: arch.DefaultLattice) -> None:
 
 
 @pytest.mark.parametrize("arc", [arch.DefaultMovement(), arch.DefaultLattice()])
-def test_illegal_gate(arc) -> None:
+def test_illegal_gate(arc):
     illegal_gate = cirq.Rx(rads=2).on(cirq.LineQubit(0))
     with pytest.raises(ValueError, match="Gate not recognized"):
         _ = arc.gate_cost(illegal_gate)
 
 
 @pytest.mark.parametrize("d", (3, 5, 7))
-def test_movement_gate_costs(d) -> None:
+def test_movement_gate_costs(d):
     # Check that all gate costs are correct for movment architectures
     arc = arch.DefaultMovement(d=d)
     qubit_a, qubit_b = cirq.GridQubit(0, 0), cirq.GridQubit(0, 1)
@@ -89,12 +86,11 @@ def test_movement_gate_costs(d) -> None:
     # Check Cultivate
 
     op = lsp.Cultivate(pi / 4).on(qubit_a)
+    cost = arc.gate_cost(op)
     if d < 7:
         with pytest.warns(UserWarning, match="Returning result for d=7"):
-            cost = arc.gate_cost(op)
             base_cost = cultivate(dsurface=d, fault_distance=3)
     else:
-        cost = arc.gate_cost(op)
         base_cost = cultivate(dsurface=d, fault_distance=3)
     expected_cost = base_cost["serial"]
     # To account for movement we add the QubitPermutationGates to the base cost
@@ -190,7 +186,7 @@ def test_movement_gate_costs(d) -> None:
 
 
 @pytest.mark.parametrize("d", (3, 5, 7))
-def test_lattice_gate_costs(d) -> None:
+def test_lattice_gate_costs(d):
     # Test that gate costs are exact for lattice architectures
 
     arc = arch.DefaultLattice(d=d)
@@ -272,6 +268,7 @@ def test_lattice_gate_costs(d) -> None:
     # Check Cultivate
 
     op = lsp.Cultivate(pi / 4).on(qubit_a)
+    cost = arc.gate_cost(op)
     if d < 7:
         with pytest.warns(UserWarning, match="Returning result for d=7"):
             cost = arc.gate_cost(op)
@@ -338,7 +335,7 @@ def test_lattice_gate_costs(d) -> None:
         assert expectation == cost
 
 
-def test_self_returns(movement_architecture, lattice_architecture) -> None:
+def test_self_returns(movement_architecture, lattice_architecture):
     # TODO: There are no self-returns anymore so this function is not well named
     qubit_a, qubit_b = cirq.GridQubit(0, 0), cirq.GridQubit(0, 1)
     ops_and_expectations = [
@@ -353,10 +350,12 @@ def test_self_returns(movement_architecture, lattice_architecture) -> None:
 
 
 @pytest.mark.parametrize("d", (3, 5, 7))
-def test_against_cultiv(d) -> None:
+def test_against_cultiv(d):
     # Test Syndrome Extract
     # Set up memory circuit
-    with open(DATA_DIR / "cultivate_costs.json") as f:
+    with open(
+        os.path.dirname(os.path.abspath(__file__)) + "/../data/cultivate_costs.json"
+    ) as f:
         saved_resources = json.load(f)
 
     d_count = load_saved_cost(dsurface=d, op_key="memory_d_rounds")["serial"]
@@ -407,7 +406,7 @@ def test_against_cultiv(d) -> None:
     assert circuit_cost[cirq.CZ] == official_cnot_resources[cirq.CZ]
 
 
-def test_movement_moment_costs(movement_architecture) -> None:
+def test_movement_moment_costs(movement_architecture):
     # Test that all primitives have moment costs
     qubit_a, qubit_b = cirq.GridQubit(0, 0), cirq.GridQubit(0, 1)
 
@@ -485,11 +484,10 @@ def test_movement_moment_costs(movement_architecture) -> None:
         _ = movement_architecture.moment_cost(op)
 
 
-def test_lattice_moment_costs(lattice_architecture) -> None:
+def test_lattice_moment_costs(lattice_architecture):
     # Test that all primitives have correct moment costs
     op = lsp.Cultivate(pi / 4).on(cirq.GridQubit(0, 0))
     cost = lattice_architecture.moment_cost(op=op)
-    pass
 
     op = cirq.H.on(cirq.GridQubit(0, 0))
     cost = lattice_architecture.moment_cost(op)
@@ -550,7 +548,7 @@ def test_lattice_moment_costs(lattice_architecture) -> None:
         _ = lattice_architecture.gate_cost(cirq.Rx(rads=7).on(cirq.GridQubit(0, 0)))
 
 
-def test_timing(movement_architecture, lattice_architecture) -> None:
+def test_timing(movement_architecture, lattice_architecture):
     # This test should break first when we introduce real gate times
     gates_with_time = [
         (cirq.PhasedXZGate, 5.0),
@@ -573,7 +571,7 @@ def test_timing(movement_architecture, lattice_architecture) -> None:
         lattice_architecture.op_time(cirq.CNOT.on(cirq.GridQubit(0, 0), cirq.GridQubit(0, 1)))
 
 
-def test_classmethods() -> None:
+def test_classmethods():
     movement_input_dict = {
         "movement": True,
         "idling": True,
@@ -670,7 +668,7 @@ def test_classmethods() -> None:
         _ = arch.Architecture.from_dict(input_dict)
 
 
-def test_dual_species_with_movement() -> None:
+def test_dual_species_with_movement():
     # HM never pays for Measurement
     # HM often pays to move for CZ
     # - Transversal CNOT
@@ -724,7 +722,7 @@ def test_dual_species_with_movement() -> None:
 
 
 @pytest.mark.parametrize("fold", (True, False))
-def test_mzo(fold) -> None:
+def test_mzo(fold):
     # MZO always pays two Moves per measure
     # - Syndrome Extract
     # - Cultiving (folded or unfolded)
@@ -778,7 +776,7 @@ def test_mzo(fold) -> None:
     assert mzo_t_cult == ssm_t_cult
 
 
-def test_string_representations() -> None:
+def test_string_representations():
     ssm = arch.DefaultMovement(
         idling=True, post_op_correction=True, d=9, cultivation_repetition=10, syndrome_rounds=1
     )
@@ -811,7 +809,7 @@ def test_string_representations() -> None:
     assert str(dsnm) == "DualSpeciesNoMovement(d=7, cr=1, fd=3)"
 
 
-def test_folded_architecture() -> None:
+def test_folded_architecture():
     folded_movement = arch.DefaultMovement(fold_cultiv=True)
     normal_movement = arch.DefaultMovement(fold_cultiv=False)
 
@@ -821,9 +819,8 @@ def test_folded_architecture() -> None:
     assert folded_cultivation_time < normal_cultivation_time
 
 
-def test_convert_globals_to_phasedxz() -> None:
-    """
-    Confirm that the conversion function works as expected
+def test_convert_globals_to_phasedxz():
+    """Confirm that the conversion function works as expected
     """
     sc = arch.Superconductor()
     example1 = {
@@ -849,7 +846,7 @@ def test_convert_globals_to_phasedxz() -> None:
     assert expected == actual
 
 
-def test_logical_move() -> None:
+def test_logical_move():
     arc = arch.DualSpeciesMovement()
     one_hop = lsp.Move(zone=None).on(cirq.GridQubit(0, 0), cirq.GridQubit(1, 0))
     two_hop = lsp.Move(zone=None).on(cirq.GridQubit(0, 0), cirq.GridQubit(1, 1))
@@ -860,7 +857,7 @@ def test_logical_move() -> None:
     assert two_cost == 2 * one_cost
 
 
-def test_y_cult_on_movement() -> None:
+def test_y_cult_on_movement():
     ssm = arch.DefaultMovement(d=11)
     mzo = arch.MeasureZonesOnly(d=11)
     dsm = arch.DualSpeciesMovement(d=11)
